requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-storage==2.18.2